% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.deriv.1st}
\alias{cheb.deriv.1st}
\title{Calculation of the values of the first derivation}
\usage{
cheb.deriv.1st(x.axis, cheb.coeff)
}
\arguments{
\item{x.axis}{beliebige X-Achse (Skalar oder Vektor)}

\item{cheb.coeff}{Chebyshev-Koeffizienten aus Least-Squares-Verfahren (Vektor)}
}
\value{
cheb.model.deriv.1st Erste Ableitung des gefilterten Modells (Skalar oder Vektor)
}
\description{
cheb.deriv.1st berechnet aus den Chebyshev-Koeffizienten die Werte der ersten Ableitung
}

